% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{detrended}
\alias{detrended}
\title{Function for class 'trend'}
\usage{
detrended(x, ...)
}
\description{
Returns the detrended signal for the trend.
}
\seealso{
\code{\link[=gettrend]{gettrend()}}, \code{\link[=getltrend]{getltrend()}}, \code{\link[=plot.trend]{plot.trend()}}, \code{\link[=time.trend]{time.trend()}}, \code{\link[=orisignal]{orisignal()}}
}
