% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrendeR.R
\docType{package}
\name{detrendeR}
\alias{detrendeR}
\title{detrendeR: This package provides a small set of functions that detect and remove linear or polynomial trend in
time-serie signal.}
\description{
The main functions are \code{\link[=getltrend]{getltrend()}} and \code{\link[=gettrend]{gettrend()}}.
}
\details{
This package does not depend on any other package or function outside of the "stat" namespace.
}
