% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getltrend}
\alias{getltrend}
\title{Returns linear trend in 'signal' time-serie}
\usage{
getltrend(signal, time = NULL)
}
\arguments{
\item{signal}{A vector or a time-serie object. Time-sorted values of signal containing trend.}

\item{time}{A positive real number. Time span of the signal in base units.}
}
\value{
Returns an object of a class "trend" containing vector of values of the estimated trend.

The object also contains:
\item{values}{original values of the signal}
\item{detrended}{the difference between the original values of the signal and the trend}
\item{time}{vector of the time domain of the signal}
}
\description{
This function detects a linear trend and returns it as a vector.
The time can be recognised if the signal is a time-serie object.
This function is a wrapper for \code{\link[=gettrend]{gettrend()}}.
}
\examples{
x <- sin(seq(0, 6, length.out=100)) + rnorm(100)
trend1 <- getltrend(x, time=1)
detrended(trend1)

data(austres)
trend2 <- getltrend(austres)
plot(trend2)
}
\seealso{
\code{\link[=gettrend]{gettrend()}}, \code{\link[=plot.trend]{plot.trend()}}, \code{\link[=time.trend]{time.trend()}}, \code{\link[=detrended]{detrended()}}, \code{\link[=orisignal]{orisignal()}}
}
