% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gettrend}
\alias{gettrend}
\title{Returns polynomial trend in 'signal' time-serie}
\usage{
gettrend(signal, degree = NULL, maxfreq = NULL, time = NULL)
}
\arguments{
\item{signal}{A vector or a time-serie object. Time-sorted values of signal containing trend.}

\item{degree}{A positive natural number. The degree of the trend polynomial.}

\item{maxfreq}{A positive real number. Maximum possible trend frequency.}

\item{time}{A positive real number. Time span of the signal in base units.}
}
\value{
Returns an object of a class "trend" containing vector of values of the estimated trend.

The object also contains:
\item{values}{original values of the signal}
\item{detrended}{the difference between the original values of the signal and the trend}
\item{time}{vector of the time domain of the signal}
\item{model}{model used to predct trend for purpose of estimating the trend on new data}
}
\description{
This function detects a polynomial trend of n-th degree and returns it as a vector.
The degree is eithe specified by the parameter 'degree' or computed from the paramters
'time' and 'maxfreq' according to the formula d = 2 time maxfreq, where maxfreq
is maximum possible frequency of the trend.
The time can be recognised if the signal is a time-serie object.
}
\examples{
x <- sin(seq(0, 6, length.out=100)) + rnorm(100)
trend1 <- gettrend(x, maxfreq=2, time=1)
detrended(trend1)
predict(trend1, newtime=seq(1, 2, length.out=100))

data(austres)
trend2 <- gettrend(austres, maxfreq=.1)
plot(trend2)
}
\seealso{
\code{\link[=getltrend]{getltrend()}}, \code{\link[=plot.trend]{plot.trend()}}, \code{\link[=time.trend]{time.trend()}}, \code{\link[=detrended]{detrended()}}, \code{\link[=orisignal]{orisignal()}}, \code{\link[=predict.trend]{predict.trend()}}
}
