% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{orisignal}
\alias{orisignal}
\title{Function for class 'trend'}
\usage{
orisignal(x, ...)
}
\description{
Returns the initial signal for the trend.
}
\seealso{
\code{\link[=gettrend]{gettrend()}}, \code{\link[=getltrend]{getltrend()}}, \code{\link[=plot.trend]{plot.trend()}}, \code{\link[=time.trend]{time.trend()}}, \code{\link[=detrended]{detrended()}}, \code{\link[=predict.trend]{predict.trend()}}
}
