% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{plot.trend}
\alias{plot.trend}
\title{Generic plot function for class 'trend'}
\usage{
\method{plot}{trend}(x, ...)
}
\description{
Plots trend detected in the signal alongside the signal.
}
\seealso{
\code{\link[=gettrend]{gettrend()}}, \code{\link[=getltrend]{getltrend()}}, \code{\link[=time.trend]{time.trend()}}, \code{\link[=detrended]{detrended()}}, \code{\link[=orisignal]{orisignal()}}
}
