% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{predict.trend}
\alias{predict.trend}
\title{Function for class 'trend'}
\usage{
\method{predict}{trend}(x, newtime, ...)
}
\description{
Predicts the trend for new time interval
Returns a 'trend' class object.
}
\seealso{
\code{\link[=gettrend]{gettrend()}}, \code{\link[=getltrend]{getltrend()}}, \code{\link[=plot.trend]{plot.trend()}}, \code{\link[=time.trend]{time.trend()}}, \code{\link[=detrended]{detrended()}}, \code{\link[=orisignal]{orisignal()}}
}
