% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{print.trend}
\alias{print.trend}
\title{Generic print function for class 'trend'}
\usage{
\method{print}{trend}(x, ...)
}
\description{
Prints trend detected in the signal.
}
\seealso{
\code{\link[=gettrend]{gettrend()}}, \code{\link[=getltrend]{getltrend()}}, \code{\link[=plot.trend]{plot.trend()}}, \code{\link[=time.trend]{time.trend()}}, \code{\link[=detrended]{detrended()}}, \code{\link[=orisignal]{orisignal()}}
}
